DJANGOCOURSE/Coffe_second_project/coffe_project/requirements-dev.txt
ipython==9.2.0
django-extensions==4.1
black